+++ pkg/fUnitRoots/src/init.c
@@ -20,4 +20,5 @@ void R_init_fUnitRoots(DllInfo *dll)
 {
     R_registerRoutines(dll, NULL, NULL, FortranEntries, NULL);
     R_useDynamicSymbols(dll, FALSE);
+    R_forceSymbols(dll, TRUE);
 }