+++ pkg/fUnitRoots/NEWS.md
@@ -9,9 +9,7 @@
 
 - in NAMESPACE, now export selectively rather than with the pattern `"."`.
 
-- set up the Fortran code to use ‘R_registerRoutines’ and ‘R_useDynamicSymbols’.
-  Haven't done this for Fortran and used an automatically generated code by
-  `tools::package_native_routine_registration_skeleton()`, which seems to work.
+- set up the Fortran code to use ‘R_registerRoutines’, etc.
   
 
 # fUnitRoots 3042.79 and older versions